+++ fluent_contents/plugins/code/models.py
@@ -1,8 +1,9 @@
 from django.db import models
-from django.utils.text import truncate_words
+from django.utils.text import Truncator
 from django.utils.translation import ugettext_lazy as _
 from fluent_contents.models import ContentItem
 from fluent_contents.plugins.code import backend, appsettings
+
 
 class CodeItem(ContentItem):
     """
@@ -18,4 +19,4 @@ class CodeItem(ContentItem):
         verbose_name_plural = _('Code snippets')
 
     def __unicode__(self):
-        return truncate_words(self.code, 20)
+        return Truncator(self.code).words(20)