+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Version 0.9
 -------------
 
+* Dropped Django 1.3 support, added Django 1.6 support.
 * Added ``FLUENT_CONTENTS_PLACEHOLDER_CONFIG`` variable to limit plugins in specific placeholder slots.
 * Added model fields for plugin developers, to have a consistent interface.
   The model fields integrate with django-any-urlfield_, django-any-imagefield_ and django-wysiwyg_.